requirements.txt
# YouTube Shorts Automation - Requirements

# Core Dependencies
requests>=2.31.0
openai>=1.12.0

# Video Processing
moviepy>=1.0.3
Pillow>=10.0.0
opencv-python>=4.8.0

# Audio Processing
pydub>=0.25.1

# YouTube Upload
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
google-auth>=2.23.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
supabase>=2.3.0

# Optional: For audio format handling (MoviePy dependency)
imageio>=2.33.0
imageio-ffmpeg>=0.4.9
